#22 Add Finalised Analysis Pipeline RQ1 and RQ2
This commit adds the finished implementation of the R analysis pipeline that runs the statistical tests, including: Friedman NHST testing, Wilcoxon pairwise post-hoc testing with Holm-Bonferroni correction, Wilcoxon effect size testing (two-sample paired signed-rank test). The combined results from the post-hoc tests are combined and saved to a csv-file for further analysis and presentation of results in a future commit.