#22 Move Web Scraping Notebook to 'Data' Folder
This commit moves the Jupyter Notebook used to web scrape the Mozilla dataset into the data folder closer to other related files such as the datasets and the 'sampling' script. The filepaths for the output files have been updated to reflect the the new location of the script, and some of the output filenames have been fixed to conform with the naming convention used for the datasets.